Initial trading rate idea
