+++ ibapi/ibkr_scanner.py
@@ -1,6 +1,5 @@
 import pandas as pd
 from ibapi.scanner import ScannerSubscription
-from ibapi.tag_value import TagValue
 from ibapi.contract import ContractDetails
 from ibkr_base import IBBase
 from time import sleep
@@ -51,9 +50,9 @@ def get_ibkr_scanner(scancode, tagvalues):
 def main():
     # Adding the required filters
     scancode = "TOP_VOLUME_RATE"
-    scancode = "HOT_BY_VOLUME"
-    scancode = "TOP_PERC_GAIN"
-    scancode = "TOP_PERC_LOSE"
+    # scancode = "HOT_BY_VOLUME"
+    # scancode = "TOP_PERC_GAIN"
+    # scancode = "TOP_PERC_LOSE"
 
     tagvalues = [
         # TagValue("priceAbove", "1"),         # Price above $1
