+++ strats/momentum.py
@@ -1,20 +1,7 @@
-from sklearn.linear_model import LinearRegression
-import numpy as np
-
-import hputils
+import strats.utils as utils
 from ibkr_data import get_ibkr_data
 from ibkr_order import IBOrder
 
-def mean_reversion_price(trade_data):
-    return sum(trade_data) / len(trade_data)
-
-def momentum_price(trade_data):
-    # using linear regression
-
-    X = np.array(range(len(trade_data))).reshape(-1, 1)
-    y = trade_data
-    model = LinearRegression().fit(X, y)
-    return model.predict([[len(trade_data)]])
 
 def main():
     trade_data = get_ibkr_data('AAPL', '', '2000 S', '5 secs')
